Returned to separate viewModels, created order is passed to fragment
